requirements.txt
Django>=5.0
psycopg2>=2.9
pillow>=10.1
djangorestframework>=3.14